requirements.txt
scenedetect==0.6.6
ffmpeg-python==0.2.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
PyYAML>=6.0
